+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 - `editor`
   - [#396](https://github.com/wix-incubator/rich-content/pull/396) sticky static toolbar on all platforms
 ### :bug: Bug Fix
+- `common`
+  - [#411](https://github.com/wix-incubator/rich-content/pull/411) base toolbar height calculated once
 - `editor`
   - [#392](https://github.com/wix-incubator/rich-content/pull/392) text inline toolbar tooltips
   - [#400](https://github.com/wix-incubator/rich-content/pull/400) inline and mobile toolbar RTL support